binary image preprocessing transform, using LineDrawingTransform

+++ training.py
@@ -1,4 +1,5 @@
 from unet import *
+from tqdm import tqdm
 
 def show_diffusion_chain(clean_image, epsilon, T):
     num_noise_levels = 20
@@ -52,8 +53,19 @@ def train(model, optimizer, loss_function, training_loader, epochs_num, device, 
             x0, _ = data
 
             x0 = x0.to(device)
-            epsilon = torch.randn(x0.shape, dtype=x0.dtype, device=device)
-            t = int(torch.randint(1, T, (1,), dtype=x0.dtype, device=device))
+            # x0 = x0.bool()
+            # x0 = x0.to(torch.bfloat16) # TODO: move to a separate case/function for binary images!
+            # show_image(x0, 'x0', block=True, cmap='gray')
+            # continue
+
+            # TODO: move to a separate case/function for binary images!
+            # epsilon = torch.randn(x0.shape, dtype=x0.dtype, device=device)
+            epsilon = torch.randn(x0.shape, device=device).to(x0.dtype)
+
+            # TODO: move to a separate case/function for binary images!
+            # t = int(torch.randint(1, T, (1,), dtype=x0.dtype, device=device))
+            t = int(torch.randint(1, T, (1,), device=device).to(x0.dtype))
+
             ts.append(t)
 
             alpha_1_to_t_array = []
